First push of working program
Program now has a config.ini to avoid exposing personal information, this is generated on first run if it doesn't already exist for the user, and has comments to show what the user should place where. It also now has a BASIC gui that will be getting improved with more functionality to avoid the need for console use at all. This will eventually get compiled into an exe for ease of use for end-user.