Major refactor in methods used
Reworked the way we were determining files, the gui look swapped from tkinter to customtkinker,, error codes added, but will need to be reworked entirely.

There are likely still errors here, and redundant functions, bad variable and function naming and missing error codes or prints. Or possibly debug prints left on. Therefore this is not a production method and further pushes will be made soon.