Attempted to make the process faster, and resolved other issues.
Attempted to make the entire update process faster by running simultaneous operations for extracting.

Updated the output messages to be clearer adding gaps where it makes sense between processes.

Updated the gui to add a stop button beside start update when the process is in operation.

Fixed the forge version grabbing and the manifest generation to now be correct again.

Fixed issue with pams not working, turns out it was from the version in solder being set incorrectly.